bahamas-tickets-bot/requirements.txt
discord.py
aiohttp
python-dotenv
aiomysql
PyMySQL
pandas
openpyxl